Move delegation-process tests to integration dir
