+++ packages/delegation-process/test/integration/getConsumerDelegators.test.ts
@@ -1,5 +1,5 @@
+/* eslint-disable @typescript-eslint/no-floating-promises */
 import {
-  getMockAgreement,
   getMockAuthData,
   getMockContext,
   getMockDelegation,
@@ -10,25 +10,23 @@ import {
   generateId,
   delegationKind,
   delegationState,
-  agreementState,
   TenantId,
 } from "pagopa-interop-models";
 import { describe, beforeEach, it, expect } from "vitest";
+
 import {
-  addOneAgreement,
   addOneDelegation,
   addOneEservice,
   addOneTenant,
   delegationService,
-} from "./utils.js";
+} from "../integrationUtils.js";
 
-describe("getConsumerDelegatorsWithAgreements", () => {
+describe("getConsumerDelegators", () => {
   const delegator1 = { ...getMockTenant(), name: "Comune di Burione" };
   const delegator2 = { ...getMockTenant(), name: "Comune di Milano" };
   const delegator3 = { ...getMockTenant(), name: "DeleganteTre" };
-  const delegator4 = { ...getMockTenant(), name: "DeleganteQuattro" };
-  const delegator5 = { ...getMockTenant(), name: "PagoPA" };
-  const delegator6 = getMockTenant();
+  const delegator4 = { ...getMockTenant(), name: "PagoPA" };
+  const delegator5 = { ...getMockTenant(), name: "DeleganteCinque" };
   const requesterId = generateId<TenantId>();
   const eservice1 = getMockEService();
   const eservice2 = getMockEService();
@@ -42,11 +40,6 @@ describe("getConsumerDelegatorsWithAgreements", () => {
     eserviceId: eservice1.id,
   });
 
-  const mockAgreement1 = {
-    ...getMockAgreement(eservice1.id, delegator1.id, agreementState.active),
-    producerId: eservice1.producerId,
-  };
-
   const mockDelegation1Bis = getMockDelegation({
     kind: delegationKind.delegatedConsumer,
     state: delegationState.active,
@@ -55,12 +48,7 @@ describe("getConsumerDelegatorsWithAgreements", () => {
     eserviceId: eservice2.id,
   });
 
-  const mockAgreement1bis = {
-    ...getMockAgreement(eservice2.id, delegator1.id, agreementState.active),
-    producerId: eservice2.producerId,
-  };
-
-  // Delegator1 has 2 active delegations and 2 active agreements
+  // Delegator1 has 2 active delegations with 2 published eservices
 
   const mockDelegation2 = getMockDelegation({
     kind: delegationKind.delegatedConsumer,
@@ -70,12 +58,7 @@ describe("getConsumerDelegatorsWithAgreements", () => {
     eserviceId: eservice1.id,
   });
 
-  const mockAgreement2 = {
-    ...getMockAgreement(eservice1.id, delegator2.id, agreementState.active),
-    producerId: eservice1.producerId,
-  };
-
-  // Delegator2 has 1 active delegation and 1 active agreement
+  // Delegator2 has 1 active delegation with 1 published agreement
 
   const mockDelegation3 = getMockDelegation({
     kind: delegationKind.delegatedConsumer,
@@ -92,45 +75,20 @@ describe("getConsumerDelegatorsWithAgreements", () => {
     state: delegationState.active,
     delegateId: requesterId,
     delegatorId: delegator4.id,
-    eserviceId: eservice1.id,
+    eserviceId: eservice3.id,
   });
 
-  const mockAgreement4 = {
-    ...getMockAgreement(eservice1.id, delegator4.id, agreementState.rejected),
-    producerId: eservice1.producerId,
-  };
-
-  // Delegator4 has 1 active delegation and 1 rejected agreement
+  // Delegator4 has 1 active delegation with 1 published eservice
 
   const mockDelegation5 = getMockDelegation({
     kind: delegationKind.delegatedConsumer,
     state: delegationState.active,
-    delegateId: requesterId,
+    delegateId: generateId<TenantId>(),
     delegatorId: delegator5.id,
     eserviceId: eservice3.id,
   });
 
-  const mockAgreement5 = {
-    ...getMockAgreement(eservice3.id, delegator5.id, agreementState.active),
-    producerId: eservice3.producerId,
-  };
-
-  // Delegator5 has 1 active delegation and 1 active agreement
-
-  const mockDelegation6 = getMockDelegation({
-    kind: delegationKind.delegatedConsumer,
-    state: delegationState.active,
-    delegateId: generateId<TenantId>(),
-    delegatorId: delegator6.id,
-    eserviceId: eservice3.id,
-  });
-
-  const mockAgreement6 = {
-    ...getMockAgreement(eservice3.id, delegator6.id, agreementState.active),
-    producerId: eservice3.producerId,
-  };
-
-  // Delegator6 has 1 active delegation and 1 active agreement but a different delegateId
+  // Delegator5 has 1 active delegation with 1 published eservice but a different delegateId
 
   beforeEach(async () => {
     await addOneEservice(eservice1);
@@ -142,52 +100,18 @@ describe("getConsumerDelegatorsWithAgreements", () => {
     await addOneDelegation(mockDelegation3);
     await addOneDelegation(mockDelegation4);
     await addOneDelegation(mockDelegation5);
-    await addOneDelegation(mockDelegation6);
     await addOneTenant(delegator1);
     await addOneTenant(delegator2);
     await addOneTenant(delegator3);
     await addOneTenant(delegator4);
     await addOneTenant(delegator5);
-    await addOneTenant(delegator6);
-    await addOneAgreement(mockAgreement1);
-    await addOneAgreement(mockAgreement1bis);
-    await addOneAgreement(mockAgreement2);
-    await addOneAgreement(mockAgreement4);
-    await addOneAgreement(mockAgreement5);
-    await addOneAgreement(mockAgreement6);
   });
 
-  it("should get delegators filtered by delegateId", async () => {
-    expect(
-      await delegationService.getConsumerDelegatorsWithAgreements(
-        {
-          offset: 0,
-          limit: 50,
-        },
-        getMockContext({ authData: getMockAuthData(requesterId) })
-      )
-    ).toEqual({
-      results: [
-        {
-          id: delegator1.id,
-          name: delegator1.name,
-        },
-        {
-          id: delegator2.id,
-          name: delegator2.name,
-        },
-        {
-          id: delegator5.id,
-          name: delegator5.name,
-        },
-      ],
-      totalCount: 3,
-    });
-  });
   it("should apply offset and limit", async () => {
     expect(
-      await delegationService.getConsumerDelegatorsWithAgreements(
+      await delegationService.getConsumerDelegators(
         {
+          eserviceIds: [],
           offset: 1,
           limit: 1,
         },
@@ -205,8 +129,9 @@ describe("getConsumerDelegatorsWithAgreements", () => {
   });
   it("should filter delegators by the 'delegatorName' parameter", async () => {
     expect(
-      await delegationService.getConsumerDelegatorsWithAgreements(
+      await delegationService.getConsumerDelegators(
         {
+          eserviceIds: [],
           offset: 0,
           limit: 50,
           delegatorName: "Comune",
@@ -228,8 +153,9 @@ describe("getConsumerDelegatorsWithAgreements", () => {
     });
 
     expect(
-      await delegationService.getConsumerDelegatorsWithAgreements(
+      await delegationService.getConsumerDelegators(
         {
+          eserviceIds: [],
           offset: 0,
           limit: 50,
           delegatorName: "PagoPA",
@@ -239,47 +165,46 @@ describe("getConsumerDelegatorsWithAgreements", () => {
     ).toEqual({
       results: [
         {
-          id: delegator5.id,
-          name: delegator5.name,
+          id: delegator4.id,
+          name: delegator4.name,
         },
       ],
       totalCount: 1,
     });
   });
-  it("should return no results if no delegations match the criteria", async () => {
+  it("should filter delegators by the 'eserviceIds' parameter", async () => {
     expect(
-      await delegationService.getConsumerDelegatorsWithAgreements(
+      await delegationService.getConsumerDelegators(
         {
+          eserviceIds: [eservice1.id, eservice2.id],
+          offset: 0,
+          limit: 50,
+        },
+        getMockContext({ authData: getMockAuthData(requesterId) })
+      )
+    ).toEqual({
+      results: [
+        {
+          id: delegator1.id,
+          name: delegator1.name,
+        },
+        {
+          id: delegator2.id,
+          name: delegator2.name,
+        },
+      ],
+      totalCount: 2,
+    });
+  });
+  it("should return an empty array if requester is not a consumer delegate", async () => {
+    expect(
+      await delegationService.getConsumerDelegators(
+        {
+          eserviceIds: [],
           offset: 0,
           limit: 50,
         },
         getMockContext({})
-      )
-    ).toEqual({
-      results: [],
-      totalCount: 0,
-    });
-
-    expect(
-      await delegationService.getConsumerDelegatorsWithAgreements(
-        {
-          offset: 0,
-          limit: 50,
-        },
-        getMockContext({ authData: getMockAuthData(delegator3.id) }) // No active delegation
-      )
-    ).toEqual({
-      results: [],
-      totalCount: 0,
-    });
-
-    expect(
-      await delegationService.getConsumerDelegatorsWithAgreements(
-        {
-          offset: 0,
-          limit: 50,
-        },
-        getMockContext({ authData: getMockAuthData(delegator4.id) }) // No active agreements
       )
     ).toEqual({
       results: [],
