+++ packages/delegation-process/test/integration/getDelegationById.test.ts
@@ -6,8 +6,8 @@ import {
   generateId,
 } from "pagopa-interop-models";
 import { describe, expect, it } from "vitest";
-import { delegationNotFound } from "../src/model/domain/errors.js";
-import { addOneDelegation, delegationService } from "./utils.js";
+import { delegationNotFound } from "../../src/model/domain/errors.js";
+import { addOneDelegation, delegationService } from "../integrationUtils.js";
 
 describe("get delegation by id", () => {
   it.each(Object.values(delegationKind))(
